Fix checkstyle violation
